Refactored Destructable.cs Changed WeaponBack.cs according designDoc Added healthbar logic small changes

+++ Assets/Resources/Scripts/IOnDamage.cs
@@ -2,6 +2,6 @@ namespace LaninCode
 {
     public interface IOnDamage
     {
-        void SetHealth(int health);
+        void SetHealth(float health);
     }
 }
